Update asset paths and add new manifest file
This commit modifies the links to icon assets in the index.html file, changes the location of the manifest file and adds new icon assets in the project.json file. It also creates a new manifest.json in the src directory with the definitions and paths for the different icon sizes.